backend/requirements.txt
openai==1.3.0
pyannote.audio==3.1.1
pydub==0.25.1
pandas==2.1.4
soundfile==0.12.1
numpy==1.24.3
torch==2.1.1
requests==2.31.0
python-dotenv==1.0.0

# Web App Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
aiofiles==23.2.1
python-docx==1.1.0 
mammoth==1.6.0 